+++ components/Stage/StageCard.tsx
