+++ components/Stage/__tests__/StageCard-test.tsx
